Factor out common integration end-to-end test components
